// src/lib/input/enums/keyboard_keycode.ts
enum KeyboardKeycode {
    BACKSPACE = 8,
    TAB = 9,
    ENTER = 13,
    SHIFT = 16,
    CTRL = 17,
    ALT = 18,
    PAUSE = 19,
    CAPSLOCK = 20,
    ESCAPE = 27,
    SPACE = 32,
    PAGEUP = 33,
    PAGEDOWN = 34,
    END = 35,
    HOME = 36,
    LEFTARROW = 37,
    UPARROW = 38,
    RIGHTARROW = 39,
    DOWNARROW = 40,
    INSERT = 45,
    DELETE = 46,

    KEY0 = 48,
    KEY1 = 49,
    KEY2 = 50,
    KEY3 = 51,
    KEY4 = 52,
    KEY5 = 53,
    KEY6 = 54,
    KEY7 = 55,
    KEY8 = 56,
    KEY9 = 57,

    KEYA = 65,
    KEYB = 66,
    KEYC = 67,
    KEYD = 68,
    KEYE = 69,
    KEYF = 70,
    KEYG = 71,
    KEYH = 72,
    KEYI = 73,
    KEYJ = 74,
    KEYK = 75,
    KEYL = 76,
    KEYM = 77,
    KEYN = 78,
    KEYO = 79,
    KEYP = 80,
    KEYQ = 81,
    KEYR = 82,
    KEYS = 83,
    KEYT = 84,
    KEYU = 85,
    KEYV = 86,
    KEYW = 87,
    KEYX = 88,
    KEYY = 89,
    KEYZ = 90,

    F1 = 112,
    F2 = 113,
    F3 = 114,
    F4 = 115,
    F5 = 116,
    F6 = 117,
    F7 = 118,
    F8 = 119,
    F9 = 120,
    F10 = 121,
    F11 = 122,
    F12 = 123,

    LEFTWINDOWS = 91,
    RIGHTWINDOWS = 92,
    SELECT = 93,

    NUMPAD0 = 96,
    NUMPAD1 = 97,
    NUMPAD2 = 98,
    NUMPAD3 = 99,
    NUMPAD4 = 100,
    NUMPAD5 = 101,
    NUMPAD6 = 102,
    NUMPAD7 = 103,
    NUMPAD8 = 104,
    NUMPAD9 = 105,

    MULTIPLY = 106,
    ADD = 107,
    SUBTRACT = 109,
    DECIMALPOINT = 110,
    DIVIDE = 111,

    NUMLOCK = 144,
    SCROLLLOCK = 145,

    SEMICOLON = 186,
    EQUALSIGN = 187,
    COMMA = 188,
    DASH = 189,
    PERIOD = 190,
    FORWARDSLASH = 191,
    GRAVEACCENT = 192,
    OPENBRACKET = 219,
    BACKSLASH = 220,
    CLOSEBRAKET = 221,
    SINGLEQUOTE = 222,
}

export default KeyboardKeycode;